moved more tests from TruModuleStandalond to TruModule tests
