+++ TruRating.TruModule.V2xx.Tests.Unit/TruRating.TruModule.V2xx.Tests.Unit.csproj
@@ -83,10 +83,10 @@
     <Compile Include="TruModuleStandaloneTests\ActiveTruModuleStandaloneTestsContext.cs" />
     <Compile Include="TruModuleStandaloneTests\WhenDoingARatingAndNotActive.cs" />
     <Compile Include="TruModuleStandaloneTests\WhenDoingARatingAndActive.cs" />
-    <Compile Include="TruModuleStandaloneTests\WhenTruModuleIsInactiveAndTTLNotExpiredAndDoingAPassiveActivationCheck.cs" />
-    <Compile Include="TruModuleStandaloneTests\WhenTruModuleIsInactiveAndTTLNotExpiredButForcingAnActivationCheck.cs" />
     <Compile Include="TruModuleTests\WhenQuestionAvailableAndDoingDwellTimeExtendRating.cs" />
     <Compile Include="TruModuleTests\TruModuleTestsContext.cs" />
+    <Compile Include="TruModuleTests\WhenTruModuleIsInactiveAndTtlNotExpiredAndDoingAPassiveActivationCheck.cs" />
+    <Compile Include="TruModuleTests\WhenTruModuleIsInactiveAndTtlNotExpiredButForcingAnActivationCheck.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config">
